Refactor babel path resolver from app to chimera.
