+++ test/_factories/field.factory.js
@@ -1,5 +1,5 @@
 import factory from 'factory-girl';
-import orm from 'app/orm';
+import orm from 'chimera/orm';
 
 const model = orm.model('ChimeraField');
